Flush the TLB after creating the initial memory mappings
bootmem_init_node() creates new mappings that override the initial sections
set up by __create_page_tables. With the addition of the supersections
support, two mappings can exist in the TLB for the same virtual address
(the original section and the newly created supersection). This is
considered undefined behaviour. The patch flushes the TLB after creating
the memory mappings for each node.

Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>