Allow the kernel to be configured for "naked" booting
The ARM Linux boot protocol is documented in Documentation/arm/Booting
but reality appears to suggest some bootloader are not Linux aware and
hoping for them to implement the outlined requirements is completely
hopeless.

Some other bootloaders are poorly maintained and getting them to provide
the required r1 value, and yet the correct one, turns out to be an
exercice in futility.

This patch is a compromize that would allow for the kernel to boot even
on those alien bootloaders with a minimalistic support: load kernel in
memory, jump to it and it is booted.

This is a configurable option so not to rule out good bootloaders, with
the help text explaining all the drawbacks that comes with it.

Signed-off-by: Nicolas Pitre <nico@cam.org>