Enable ARM OProfile support for PMU less platforms
The current code errors out during compile for platforms that do
not have PMU cycle counter. This patch allows platforms such as
the ARM Versatile to be able to use the OProfile via the CPU
Timer Interrupt option. The arch oprofile init function will fail
and allow the timer interrupt init function to be called.

Signed-off-by: Dave Jiang <djiang@mvista.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>