Fix ioremap_cached()/ioremap_wc() for SMP platforms
Write combining/cached device mappings are not setting the shared bit,
which could potentially cause problems on SMP systems since the cache
lines won't participate in the cache coherency protocol.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
--
This doesn't make it any more legal to use ioremap*() on main system
RAM than it was before this change; it does mean that ioremap_wc()
regions are correctly marked as shared which happens to agree with
the main system RAM mappings.  Strictly to the arch. spec. for ARMv6+,
it remains "unpredictable" to use ioremap*() on main system RAM.

 arch/arm/mm/mmu.c |    4 ++++
 1 files changed, 4 insertions(+), 0 deletions(-)
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>