ARM: PL310: support 16-way associativity in l2x0_inv_all()
The function to invalidate all ways in the L2 cache controller
currently assumes a configuration of 8 ways or less. It is
permitted to build a PL310 with support for 16-way associativity,
although I'm not aware of any current hardware with this feature.

This patch adds a new config option, CACHE_PL310_ASSOC_16_WAY
that can be checked by any per-way L2 cache operations in order
to invalid all of the relevant data. Currently the only function
that uses per-way operations is l2x0_inv_all(), which this patch
fixes.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>