ARM: Realview/Versatile: extend sched_clock range to improve load balancing
The old definition of sched_clock() for Realview and Versatile boards
has a maximum value of 179s, after which it overflows. This limitation
has been observed to cause load balancing issues in an SMP environment,
where one of the cores will remain idle despite an abundance of busy
tasks on the system.

This patch ports the sched_clock() implementation from RMK's
versatile branch, which has a maximum value of about 35583 days.
Basic tests have shown this drastically improves the load balancing of
the scheduler.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>