ARM: Vexpress: don't use arch_reset() to restart board
Performing a restart or poweroff operation on the Versatile Express
board requires the ability to access the System Register Interface
in order to tell the microcontroller what we want to do.

If we try to do this using arch_reset(), memory will have been
remapped as 1:1 to allow a soft-boot to occur. This means we will
get stuck in an infinite loop polling a completion bit on the
micro.

Since we don't care about soft-boot, we re-assign the pm_power_off
and arm_pm_restart to call the board functions directly which allows
the microcontroller to handle the power operations.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>