ARM: Vexpress: no-mmu: don't use HIGH_PHYS_OFFSET addresses in decompressor
For MMU-less builds, the decompressor needs to be given usable
addresses describing whereabouts it lives. Currently, these
are hardcoded to exist in the DDR2 memory at 0x60000000, but
this is inaccesible without an MMU.

This patch uses low addresses for the Kernel load address and
entry point unless HIGH_PHYS_OFFSET is enabled in the config.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>