RT62657 Support passing of the command line from U-boot.
This commit makes changes to the kernel to support passing
a kernel command line from U-boot (rather than hardcoding
kernel parameters in the build-time command line & hard-coded
tags.

Also, U-boot now passes a MACHID for the A2F.

Also, removed dropbear from initramfs. SSH is not a part of this
release so it shouldn't be included in the demo.

Also, added a device node for ttyS0 into initramfs. We open
sh on this device (to allow ^c and other control-temrinal features).