RT72064. stm32 uart: move register map to *.c file
Instead of using the real end of the resource, let's specify just
the .end = .start + 1. And I think that's absolutely OK: the driver
needs to know only the start address of the register block. If for
some reason, on some platform, the size (and content) of the register
block differs from the one assumed in driver - there we won't be able
to work: regardless the fact is .end of resource correct, or not.

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>