RT73025. ea-lpc1788: configure timers (clockevents, clocksource)
TIMER0 is used for clockevents. Match channel 0 (MR0) is used to set
the frequency of timer interrupts to CONFIG_HZ.

TIMER1 is used for clocksource. For this timer, the prescale register is
set to 0xFFFFFFFF in order to get a 64-bit timer counter:
1. The most-significant word of the 64-bit counter can be read from
       the Timer Counter Register
2. The least-significant word of the 64-bit counter can be read from
       the Prescale Counter Register.

This patch also adds the `lpc178x_periph_enable()` power management
function into the Linux kernel.