RT73025. ea-lpc1788: fix hang-up in Ethernet driver after SYSRESET
To avoid hang-up of the Ethernet block after Cortex-M3 software reset
(SYSRESET), we need to reset the Ethernet PHY immediately before
performing the SYSRESET.

If Ethernet is not enabled in the Linux kernel configuration,
we do not perform the PHY reset.
This leads to a minor bug: if you run Linux without Ethernet
support and do a `reboot`, the Ethernet driver will hang either in
Linux (if you do not use Ethernet in U-Boot) or in U-Boot.
The hang-up can be however avoided by doing a full reset (using for
example the SW1 button) right after running `reboot` in Linux.