RT75957. twr-k70f120m: configure timers (clockevents, clocksource)
PIT (Periodic Interrupt Timer) channel 0 is used for clockevents.
The frequency of timer interrupts is set to CONFIG_HZ.

The Cortex-M SysTick timer is used for clocksource. The is no better
timers that could be configured more easily.

This patch also adds the `kinetis_periph_enable()` power management
function into the Linux kernel.