RT75957. twr-k70f120m: reuse the `fec.c` Ethernet driver for Kinetis MCU
 * Use the htonl/ntohl/htons/ntohs functions when working with Ethernet
DMA descriptors. This is required for Kinetis, but the same code will
still work on other targets since they are big-endian.
 * Initialize the rate of the Ethernet clock (CLOCK_MACCLK). This is
necessary to calculate the clock divider for the MDIO interface.
 * Limit the MDIO clock to 800 kHz.
 * Pass the MAC clock rate and the MDIO clock rate limit from the
platform-specific code to the FEC driver.
 * Freescale Kinetis provides 3 standard Ethernet interrupts (Tx, Rx,
error/MDIO), all of them will be handled with the same IRQ handler
function.
 * The `fec.c` driver did not support the Micrel KSZ8041NL PHY device.
We add support for it. This PHY device is compatible with KS8721BL, so
we reuse the code already written for KS8721BL.
 * Add support for the RMII mode. This is required, because only the
RMII mode is possible on the TWR-K70F120M board.
 * In order to choose the MAC address, firstly parse the `ethaddr=`
kernel parameter. If there is no such parameter in the kernel bootargs,
generate a random MAC address. (I.e. we use the same approach as we did
for other Cortex-M targets - STM32F2 and LPC178X.)
 * Move `drivers/net/fec.h` to `include/linux/fec.h`. This is necessary,
because we want to define `struct fec_platform_data` in this header
file and this header file should be visible from both the
`drivers/net/fec.c` driver and the platform-specific code in e.g.
`arch/arm/mach-kinetis/eth.c`.

Why there are still some `#ifdef CONFIG_ARCH_KINETIS` preprocessor
directives left in the FEC driver:
1. We set FEC_ALIGNMENT to the correct value for Kinetis in the FEC
driver, because we cannot get rid of the `#ifdef CONFIG_ARCH_KINETIS`
for choosing FEC_ALIGNMENT unless we update the platform-specific code
for a bunch of platforms that use the FEC driver (including ColdFire
platforms.)
2. We use our custom `fec_get_mac()` only for Kinetis, because we
do not want to accidentally break something on other platforms.
3. In our Linux port for Kinetis, we do not use the `clk_*`
infrastructure for controlling various clocks. That is why we do not
even want to build the implementation of that infrastructure.
By the "`clk_*` infrastructure" I mean the following API functions:
   * `clk_get()`
   * `clk_put()`
   * `clk_enable()`
   * `clk_disable()`
   * `clk_get_rate()`
4. On Kinetis, we must initialize the MAC address in the hardware
manually, so we call `fec_update_mac_address_hw()` from `fec_restart()`
only on Kinetis.