RT77452. Add support for Emcraft's LPC-LNX-EVB board
 * Update `arch/arm/mach-lpc178x/` to support the new board along with
the existing EA-LPC1788 board from a single kernel. The boards are
distinguished at run-time based on `lpc178x_platform=` kernel
parameter passed by U-Boot to the kernel.
 * Add cases for `PLATFORM_LPC178X_LNX_EVB` in the following files:
  * `clock.c` - To select the correct frequency of the on-board
oscillator.
  * `eth.c` - To pass the correct PHY ID mask to the Ethernet driver and
to inform the Ethernet driver that we would like to use automatic
PHY discovery on this board.
  * `fb.c` - To skip initialization of the LCD controller on this board.
Do not register the framebuffer driver for now, because we do not
know what types of LCD screens will be connected to our board.
  * `flash.c` - To set the correct size of the JFFS2 partition in the
NOR Flash.
  * `i2c-gpio.c` - To disable registration of the I2C bus for this
board. We do not register the I2C buses with the `i2c-gpio` driver
since we do not use I2C devices on LPC-LNX-EVB yet.
  * `iomux.c` - To select the custom IOMUX configuration table for this
board. Do not configure the SD Card and USB pins for LPC-LNX-EVB,
these interfaces are not supported on this board.
  * `sdcard.c` - To skip initialization of the SD card controller which
cannot be used on this board.
 * Do not get rid of call to `lpc178x_phy_final_reset()` right before
soft-resetting the board, because it is non-trivial to make the
caller code board-specific.