RT77452. mtd: cfi_cmdset_0002: Tone down warning messages about TopBottom CFI field
Accept values of 2-5 for TopBottom, where the newly-added 4 and 5 values
mean a uniform layout. It does indicate WP layout but we don't handle that.

Also don't say "broken" when swapping erase regions in a top-boot chip.
That got retrospectively documented in the spec.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>

Conflicts:

	drivers/mtd/chips/cfi_cmdset_0002.c

The SHA1 hash of the original patch is 412da2f6e083eba6e4bd91ff2e78abb4735357a7