RT78702. stm32_usart: Reset offsets in DMA buffers on reinitialization
After reinitialization performed in `stm_port_startup()`, the DMA
controller starts writing new data to the beginning of the first
buffer, therefore the buffer indices and offsets should be reset to 0
again. Otherwise we will be reading data from wrong positions in the
buffer.

`stm_port_startup()` is called on every `open()` system call, this is
why the bug described above manifests after the second `open()` call to
open one of the TTYs.