RT78855. crtouch_mt: Make changes to support touchscreen of TWR-LCD-RGB with Qt Embedded
 * Use preprocessor variables to define what features of the touchscreen
driver should be supported.
 * Per-K70 (#ifdef CONFIG_ARCH_KINETIS used), disable the following
features of the touchscreen driver:
   * Multi-touch and gestures. Qt does not support multi-touch events
coming from a Linux input event device (/dev/input/event0), Qt
writes error messages in reply to these events.
   * Capacitive touch pads. We do not have support for them in the Linux
user-space. The capacitive touch pads are disabled by default in
the CRTouch module on the TWR-LCD-RGB board.
   * Reading of touchscreen resolution from CRTouch's I2C registers (see
function read_resolution() in crtouch_mt.c), because when trying to
read them we always get "NAK bailout" errors when using the
i2c-gpio I2C host driver.
   * Handling of IRQs from the touchscreen. We also get "NAK bailout"
error when trying to use the IRQ signal of the touchscreen.
   * Going to suspend state (crtouch_suspend(), crtouch_resume()). The
WAKE signal is not connected on the TWR-LCD-RGB board.
 * Set the IRQ polling period to 20ms (50Hz). The IRQ signals from the
touchscreen are sent from TWR-LCD-RGB also at 50Hz when the screen is
touched, so we guarantee the same touchscreen event latency as when
"real" IRQ handling is used.