RT76540. serial: kinetis_uart: Pass correct pointer into request_irq()
In the IRQ handler we cast `dev_id` to `struct kinetis_uart_priv *`,
therefore we must pass this kind of a pointer.

The old code worked because `port` is the first member of
`struct kinetis_uart_priv`, and therefore a pointer to `port` is the
same address as a pointer to the whole structure.