RT79374. kinetis ehci: Do not let CPU hang if clock is missing on ULPI_CLK
 * It appears that the MCU hangs on writing to USBHS_PORTSC1 if there is
no input clock on ULPI_CLK. This problem is not described in the Kinetis
K70 errata sheet, but it (or a very similar issue) is described in the
errata sheet for MPC8349E:
    USB24: A write to some USB registers such as PORTSC can cause
    the system to hang when no USB PHY clock is applied.
 * The problem of missing clock on ULPI_CLK may happen if you forgot to
plug a TWR-SER2 board into the Tower System.
 * In order to determine if there is a clock on the ULPI_CLK pin, we
temporarily configure it as a General Purpose Input and watch if it
switches from low level to high level and forth. If it does, then we
most likely have a clock on that input and therefore can safely start
using the USB High Speed controller.
If the level on the ULPI_CLK input looks stable, we print an error
message.