RT80044. kinetis fb: Autoconfigure LCDC clock divider to get precise LCD clock rate
 * We need more precise LCD clock rates, because of SDRAM performance
issues when driving a 800x480 display with a 30MHz pixel clock. The
correct clock rate for such a display is 23.8MHz, and the SDRAM
performance issue eliminate at this clock rate.
 * The "imxfb" framebuffer driver does not call clk_set_rate() to change
the LCD pixel clock frequency. Instead, it writes to the PCR register
to set the simple clock divider integrated in the LCD module of
Kinetis K70. This divider does not support fractional divide values.
Since we want to make the clock rate very precise, we have to use the
LCD fractional divider configurable via the SIM_CLKDIV3 register.
 * Configure the LCD clock fractional divider in
arch/arm/mach-kinetis/fb.c, after we know the requested pixel clock
rate from imx_fb_videomode[0].mode.pixclock.
 * Calculate numerator and denominator for the fractional divider using
Diophantine approximation in order to get as close as possible to the
requested pixel clock.
 * Use a straight-forward implementation of Diophantine approximation:
    1. Store the integer parts of the continued fraction in an array;
    2. At each iteration, calculate the intermediate approximation.
    3. Stop iterations when numeration or denominator exceeds the
respective limits defined by the bit fields of the LCDC clock divider in
the SIM_CLKDIV3 register (LCDCDIV is limited to 4096, LCDCFRAC is
limited to 256).
 * Avoid usage of true floating-point numbers, instead we use fixed-point
numbers implemented as unsigned 32-bit integers divided by 2**16.
 * Length limit for the continued fraction in our algorithm is the
length of the array where we store integer parts of the continued
fraction. In the worst case of the integer parts in the continued
fraction being all ones (golden ratio = [1; 1, 1, 1, ...]), numerator
and denominator of the evaluated continued fraction grow like Fibonacci
numbers, or like powers of golden ratio. This is why for the array
length we choose the logarithm of 4096 (LCDCDIV maximum value) to base
of golden ratio (1.618...), plus a few extra array elements for safety.
 * After changing the LCDC clock divider values, update the actual
clock rate in the "rate" property of the LCDC clock registered with the
clock driver, so that the imx-fb driver know the real LCDC clock
frequency and does not try to increase the internal clock divider in the
PCR register.