RT79979. rtc: Add STM32F2 RTC driver
 * STM32F1 RTC is not compatible.
 * The STM32F2 RTC module requires unlocking the RTC write protection
before writing to the majority of its registers, for example when
changing the current time. We implement a convenience function
stm32f2_rtc_write_enable() to do this job.
 * The STM32F2 RTC module requires switching to the so-called
Initialization Mode before changing the current time. We implement a
convenience function stm32f2_rtc_enable_init_mode() to do this job.
 * We will use the 24-hour notation for time, not the AM/PM notation
also supported by the RTC of STM32F2.
 * The RTC on STM32F2 stores only the last two decimal digits of the
year, i.e. there is no register field for the current century. When
reading the current time, we assume the year to be between 2001 and
2099.
 * The RTC on STM32F2 codes Sunday as a day of week number 7. This
contradicts with the Linux kernel, where Sunday is the day of week
number 0. We do necessary forward and backward conversion of
day-of-week in read_time and set_time.
 * Implement Linux RTC alarm via Alarm A.
 * Emulate Update Interrupts using STM32F2 RTC Alarm B.
 * Implement Periodic Interrupts using STM32F2 RTC wakeup timer.
 * No support for the "rtc_class_ops.proc" RTC class callback.