RT88844: Develop an SPI device driver for the uClinux Kinetis
This patch ports the SPI device driver from the Vybrid Linux kernel to Kinetis
uClinux. The target platform is the K70 SOM plugged into a SOM-BSB-EXT.

The following configuration options were added:
- KINETIS_SPI0: enable SPI controller 0
- KINETIS_SPI1: enable SPI controller 1
- KINETIS_SPI2: enable SPI controller 2
- SPI_KINETIS: enable using of selected controllers in master mode

The source files used in that patch can be found in the vybrid kernel here:
  - arch/arm/plat-mxc/include/mach/spi-mvf.h
  - drivers/spi/spi_mvf_dspi.c

This patch doesn't add DMA support for the controller.