ARM: nommu: fix warning with checksyscalls.sh
CALL    scripts/checksyscalls.sh
<stdin>:46:1: warning: "__IGNORE_migrate_pages" redefined
In file included from <stdin>:2:
arch/arm/include/asm/unistd.h:482:1: warning: this is the location of the previous definition

This is caused because we define __IGNORE_migrate_pages to be 1, but in
the case of nommu, it's defined to be empty.  Fix this by just defining
the __IGNORE_ symbols to be empty.