RT #103951. Set the value of a GPIO pin before configuring it as output
In the Linux GPIO API there is a possibility to configure a GPIO pin as
output with the value 1 using the "high" keyword instead of the "out":
  # echo 100 > /sys/class/gpio/export
  # echo high > /sys/class/gpio100/direction
However, in the current Kinetis implementation this API is not working,
because of the value of the GPIO pin is set after its configuration as
output. Thus, the GPIO pin 100 in the example above will be configured
as output, set to 0, then set to 1, even if it was set as 1 in the HW.
This behavior leads, e.g., to LCD blink when configuring LCD backlight
enable pin. The patch fixes this bug.